product crud function added with response

+++ src/main/java/com/infinity/springrestapi/repositories/CategoryRepository.java
@@ -3,5 +3,5 @@ package com.infinity.springrestapi.repositories;
 import com.infinity.springrestapi.entities.Category;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategoryRepository extends CrudRepository<Category, Long> {
+public interface CategoryRepository extends CrudRepository<Category, Byte> {
 }
